netlink: use uint32_t instead of size_t for attribute length
According to libnftables change 437d610, now the length obtained
via getter function is uint32_t, not size_t anymore.

+++ include/netlink.h
@@ -34,14 +34,14 @@ extern struct nft_rule_expr *alloc_nft_expr(const char *name);
 extern struct nft_set *alloc_nft_set(const struct handle *h);
 
 struct nft_data_linearize {
-	size_t		len;
+	uint32_t	len;
 	uint32_t	value[4];
 	char		chain[NFT_CHAIN_MAXNAMELEN];
 	int		verdict;
 };
 
 struct nft_data_delinearize {
-	size_t		len;
+	uint32_t	len;
 	const uint32_t	*value;
 	const char	*chain;
 	int		verdict;
